KREST-10313 Prototype of batch REST Produce

+++ kafka-rest/src/test/java/io/confluent/kafkarest/resources/v3/ProduceActionTest.java
@@ -815,7 +815,7 @@ public class ProduceActionTest {
         .setClusterId("clusterId")
         .setTopicName("topicName")
         .setPartitionId(partitionId)
-        .setOffset(offset)
+        .setOffset((long) offset)
         .setTimestamp(Instant.ofEpochMilli(0))
         .setErrorCode(HttpStatus.OK_200)
         .build();
